requirements.txt
openai
python-dotenv
langchain
chromadb
requests
tiktoken